Moved utility methods to util.ts, but commented them out for now. Not sure we need to keep them.
