+++ src/util.ts
@@ -104,3 +104,66 @@ export function normalizeAngle(a: number): number {
 export function sumArray(arr: number[]): number {
   return arr.reduce((a, b) => a + b, 0);
 }
+
+//////////////////////////////////////////////////////////////////////////////////////////////////
+// Moved from vex.ts to util.ts.
+
+/**
+ * Take `arr` and return a new array consisting of the sorted, unique'd,
+ * contents of arr. Does not modify `arr`.
+ */
+// eslint-disable-next-line
+/*
+  static sortAndUnique(arr: any[], cmp: any, eq: any): any[] {
+    if (arr.length > 1) {
+      const newArr = [];
+      let last;
+      arr.sort(cmp);
+
+      for (let i = 0; i < arr.length; ++i) {
+        if (i === 0 || !eq(arr[i], last)) {
+          newArr.push(arr[i]);
+        }
+        last = arr[i];
+      }
+
+      return newArr;
+    } else {
+      return [].concat(arr); // 
+    }
+  }
+  */
+
+// Get the 2D Canvas context from DOM element with id `canvasId`.
+/*
+  static getCanvasContext(canvasId: string): RenderingContext {
+    if (!canvasId) {
+      throw new RuntimeError('BadArgument', 'Invalid canvas selector: ' + canvasId);
+    }
+
+    const canvas = document.getElementById(canvasId) as HTMLCanvasElement;
+    if (!(canvas && canvas.getContext)) {
+      throw new RuntimeError('UnsupportedBrowserError', 'This browser does not support HTML5 Canvas');
+    }
+
+    return canvas.getContext('2d') as RenderingContext;
+  }
+  */
+
+/** Benchmark. Run function `f` once and report time elapsed shifted by `s` milliseconds. */
+/*
+  static benchmark(s: any, f: any): void {
+    const startTime = new Date().getTime();
+    f();
+    const elapsed = new Date().getTime() - startTime;
+    log(s, elapsed + 'ms');
+  }
+  */
+
+// Get stack trace.
+/*
+  static stackTrace(): string | undefined {
+    const err = new Error();
+    return err.stack;
+  }
+  */
