+++ src/version.ts
@@ -1,3 +1,4 @@
+// Gruntfile.js uses string-replace-loader to replace these values during build time.
 export const VERSION: string = '__VF_VERSION__';
 export const ID: string = '__VF_GIT_COMMIT_ID__';
 export const DATE: string = '__VF_BUILD_DATE__';
